Modifications for compatibility with mikroC

+++ sha2.h
@@ -52,17 +52,17 @@ typedef struct _SHA512_CTX {
 } SHA512_CTX;
 
 void SHA256_Init(SHA256_CTX *);
-void SHA256_Update(SHA256_CTX*, const uint8_t*, size_t);
+void SHA256_Update(SHA256_CTX*, const uint8_t*, uint32_t);
 void SHA256_Final(uint8_t[SHA256_DIGEST_LENGTH], SHA256_CTX*);
 char* SHA256_End(SHA256_CTX*, char[SHA256_DIGEST_STRING_LENGTH]);
-void SHA256_Raw(const uint8_t*, size_t, uint8_t[SHA256_DIGEST_LENGTH]);
-char* SHA256_Data(const uint8_t*, size_t, char[SHA256_DIGEST_STRING_LENGTH]);
+void SHA256_Raw(const uint8_t*, uint32_t, uint8_t[SHA256_DIGEST_LENGTH]);
+char* SHA256_Data(const uint8_t*, uint32_t, char[SHA256_DIGEST_STRING_LENGTH]);
 
 void SHA512_Init(SHA512_CTX*);
-void SHA512_Update(SHA512_CTX*, const uint8_t*, size_t);
+void SHA512_Update(SHA512_CTX*, const uint8_t*, uint32_t);
 void SHA512_Final(uint8_t[SHA512_DIGEST_LENGTH], SHA512_CTX*);
 char* SHA512_End(SHA512_CTX*, char[SHA512_DIGEST_STRING_LENGTH]);
-void SHA512_Raw(const uint8_t*, size_t, uint8_t[SHA512_DIGEST_LENGTH]);
-char* SHA512_Data(const uint8_t*, size_t, char[SHA512_DIGEST_STRING_LENGTH]);
+void SHA512_Raw(const uint8_t*, uint32_t, uint8_t[SHA512_DIGEST_LENGTH]);
+char* SHA512_Data(const uint8_t*, uint32_t, char[SHA512_DIGEST_STRING_LENGTH]);
 
 #endif
